t.g1.15.03.1
Security Updates
- Drupal core 7.35
- Entity API 7.x-1.6
- Webform 7.x-4.5
- Chaos tool suite (ctools) 7.x-1.7

Patched - Password Policy Module
- Patch - https://www.drupal.org/node/2426317
- Patch - https://www.drupal.org/node/2372935

The Password Policy module was added on the last update. It is young in age at version 1.10, just out of beta. There are other modules available but this one is the best solution for user friendliness and to meet the requirements of the State. There are some issues that will be resolved in the next released version. In the meantime, we had to apply patches to the current version to resolve issues on live sites.